Test for search form with change field
